resolver: extract peer-context pass into its own module
lib.rs was 6393 lines with the BFS algorithm, peer-context pass, and
a pile of general helpers all in one file. Move the peer-dependency
post-processing — `detect_unmet_peers`, `hoist_auto_installed_peers`,
`PeerContextOptions`, `apply_peer_contexts`, and their private helpers
(`visit_peer_context`, `dedupe_peer_variants`, `dedupe_peer_suffixes`,
`hash_peer_suffix`, etc.) — into `peer_context.rs`.

Tests stay in lib.rs. The private helpers they poke (`dedupe_peer_variants`,
`dedupe_peer_suffixes`, `hash_peer_suffix`, `apply_dedupe_peers_to_key`,
`contains_canonical_back_ref`) are now `pub(crate)` in peer_context and
reimported with `#[cfg(test)]` in lib.rs. `version_satisfies` moves from
private to `pub(crate)` so peer_context can call it.

Public API unchanged: `aube_resolver::UnmetPeer` and friends still
resolve via `pub use`. lib.rs drops from 6393 to 5146 lines; peer_context
is ~1280 lines. No behavior change.
